Snake Game Python
About Snake Game Python Project
The objective of this python project is to build a snake game project. In this python project, the player has to move a snake so it touches the fruit. If the snake touches itself or the border of the game then the game will over.


Project Prerequisites
To build the snake game project we used the turtle module, random module, time module, and concept of python.

Turtle module gives us a feature to draw on a drawing board

A random module will be used to generate random numbers

Time module is an inbuilt module in python. It provides the functionality of time.